Extract Setup class.
